requirements.txt
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.12.0
fastapi>=0.100.0
uvicorn>=0.23.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
pywavelets>=1.4.0
diffusers>=0.21.0
boto3>=1.28.0
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
parquet>=1.0.0
pyarrow>=12.0.0
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
python-multipart>=0.0.6
